requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
marshmallow==3.20.1
python-dotenv==1.0.0
flask-cors==4.0.0
flask-swagger-ui==4.11.1
requests==2.31.0
